Add recordsLock mutex to prevent data race for records state
Previously, p.Records was written to and read from by different
goroutines in some usages of this package. For example, this race
would be hit if we have a web UI that sorts Probes, which calls
p.Less(), which reads p.Records(), while another goroutine runs the
probe with p.Probe(). Since there was no mechanism to coordinate
access, this was a data race, as found by running a binary built with
`go build -race`. This change makes p.records internal, instead adding
a p.Records() method to access the records after acquiring the RWMutex
lock.