Guard reads/writes to p.badness with alertLock
Previously, p.Badness was written to and read from by different
goroutines in some usages of this package. For example, this race
would be hit if we have a web UI that sorts Probes, which calls
p.Less(), which reads p.Badness, while another goroutine runs the
probe with p.Probe(). Since there was no mechanism to coordinate
access, this was a data race, as found by running a binary built with
`go build -race`. This change makes p.badness internal, instead adding
a p.Badness() method to access the current badness value, after
acquiring the RWMutex lock.